backend/requirements.txt
amqp==2.2.2
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.5
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-celery-beat==1.1.1
django-cors-headers==2.2.0
django-filter==1.1.0
django-suit==0.2.26
djangorestframework==3.8.0
gprof2dot==2016.10.13
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pycountry==18.2.23
Pygments==2.2.0
PyJWT==1.6.1
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
twilio==6.12.1
Twisted==18.4.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
xlwt==1.3.0
zope.interface==4.5.0

# need for WebSocket
autobahn==18.6.1
channels==2.1.2
daphne==2.2.0
twisted==18.4.0
txaio==2.10.0
aioredis==1.1.0
channels-redis==2.2.1
hiredis==0.2.0
msgpack==0.5.6


# test DB request tool. We need to check requests speed in home page
django-silk==3.0.1